as we have been treading on each others toes I've moved my versions of the example app to demo2 and demo3 and create a separate directory with the utils classes used by both of these.
