+++ example/lib/demo_util/demo_common.dart
@@ -4,8 +4,8 @@ import 'dart:typed_data';
 import 'package:flutter_sound/flutter_sound.dart';
 import 'package:intl/intl.dart';
 
-import 'media_path.dart';
-import 'util/log.dart';
+import 'demo_media_path.dart';
+import '../util/log.dart';
 
 /// Describes how the media is stored.
 enum MediaStorage {
