+++ example/lib/demo_util/remote_player.dart
@@ -4,7 +4,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_sound/flutter_sound.dart';
 
-import 'active_codec.dart';
+import 'demo_active_codec.dart';
 
 /// path to remote auido file.
 const String exampleAudioFilePath =
